+++ changelog.md
@@ -1,2 +1,2 @@
 ### Pottery 1.0.2
-- Initial release of Pottery for Minecraft 1.21.6 & 1.21.7
+- Initial release of Pottery for Minecraft 1.21.9 & 1.21.10